Refactor: Net6 Example namespace
Refactor: Net6 Example namespace
Refactor: Update README.md
