+++ MoesifNet6Example/Properties/launchSettings.json
@@ -15,7 +15,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       }
     },
-    "MoesifNet5Example": {
+    "MoesifNet6Example": {
       "commandName": "Project",
       "launchBrowser": true,
       "environmentVariables": {
